test(delivery): replace StartTransitUseCase tests with PickUpDeliveryUseCase tests
Replace start-transit.spec.ts with pick-up-delivery.spec.ts. Remove courier repository and courier creation from test setup. Update test cases to verify delivery pickup without courierId parameter and ASSIGNED status instead of WAITING_PICKUP. Remove courier existence validation test and update error assertions to use CannotPickUpDeliveryError
